requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.43.0

# Inference
llama-cpp-python>=0.2.56

# Demo
gradio>=4.20.0

# Eval
sentencepiece
protobuf
